Handle Jenkins not launching a job.
If Jenkins fails to launch a job, a change can get stuck in the
queue.  This is very similar to when we lose track of a build in
Jenkins, so in this case, pretend the build was launched, and mark
it as LOST.  The queue will eventually drop the change and things
will keep moving.

Also, Jenkins sometimes erroneously reports 404 on the build URL.
If we get an error launching a job, just retry it for 30 seconds
before we finally give up.

Fixes bug #1011907.

Change-Id: I605fe37abf2fc3e7685df0f7a8e460b9a5d508b3