Keep build information around longer.
Preparing for leaving informative messages on build descriptions.

I'd like to be able to leave descriptions that include all the
builds that happened in conjunction with any given build.  But
Zuul forgets about completed builds very quickly, so we may not
have all the information when needed.  This adds build sets,
which keep track of all the builds that happened at once.

This makes some of our recent status output changes a little
easier/cleaner as well.

This also adds the jenkins method to set a build description,
which is unused in this change, but will be in a follow-on.

Change-Id: I78936dc9284ef0a88272ffdef962718b20485ec7