Add parameter function and job branch specs.
Add a parameter-function attribute to jobs that specifies a function
that should be called to manipulate parameters passed to jenkins
before a job is launched.

Add the ability to include a python file to define such a function.

Finish implementing the "branch" attribute of jobs that lets you
specify whether a job should run on a particular branch.

Change-Id: I3f4d21ad5ac58a24d44a9a8437daa5c668967db9