Just check that ref has changed when merging.
Record the sha at refs/head/branch before each report.  Then if
isMerged is called, just check that the sha has _changed_ since
then, and if it has, assume that means it has reached the replica.

This handles the case of a merge or cherry-pick commit being the
actual commit that ends up at the ref, without the overhead of
fetching and parsing that commit to see if it's really the change.
This should be sufficient as long as Zuul is the only thing
triggering merges.

Change-Id: I50cfa7b4c011d9fd108046914c7bd236858ff5d3
Reviewed-on: https://review.openstack.org/10970
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Approved: James E. Blair <corvus@inaugust.com>
Tested-by: Jenkins