Fix test race condition.
The fake jenkins queue may have been modified during the call to
get_queue_info, so operate on a copy of the queue (which is more
like reality than locking the queue because we get a copy of the
queue from the jenkins api).

Also, fix a string formatting error which for some reason was
not raising an exception until we tried running the tests on
quantal.

Change-Id: I51e588fb147515c826395439516cf843c2da83f3
Reviewed-on: https://review.openstack.org/23781
Reviewed-by: Jeremy Stanley <fungi@yuggoth.org>
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Approved: Jeremy Stanley <fungi@yuggoth.org>
Tested-by: Jenkins