Report an unknown time if a change ahead has failed
We don't actually know how long it will take for a change to report
if a change ahead of it has failed (because this may need to be
restarted).  We could do a better job of estimating a minimum if
we had historic run-times of jobs internally, but we don't right
now, so at least let's say it's unknown for now.

Change-Id: I7a3e4d07a02893122564817f8119cb54632464f4