Add a graphite key for all jobs in a pipeline
We have a graph on our status page showing all of the jobs Zuul
launched, but it's built from more than 1000 graphite keys which
is a little inefficient.  Add a key for convenience that rolls
up all of the job completions in a pipeline, so that such a graph
can be built with only about 10 keys.

Change-Id: Ie6dbcca68c8a118653effe90952c7921a9de9ad1