Set merge result correctly on exceptions
If an exception was received during a report, _reportItem would
erroneously indicate that it had been reported without error.
If a merge was expected, isMerged would be called which may then
raise a further exception which would stop queue processing.

Instead, set the default return value for _reportItem to True
because trigger.report returns a true value on error.  This will
cause the change to be marked as reported (with a value of ERROR),
the merge check skipped, and the change will be quickly removed
from the pipeline.

Change-Id: I08b7cee486111200ac9857644d478727c635908d