Allow multiple invocations of the same job
Previously if a job was listed more than once for a project, it was
ignored.  That's pretty arbitrary (it silently dropped the second
without an error; and if there are two, which is the right one
anyway?).  OTOH, it's potentially useful to run a job more than
once in order to increase the chance of encountering
notdeterministic behavior.  And if listing a job twice is an error,
it is now more likely to be noticed by the operator.

This removes the check for duplicate invocations of a job.

Change-Id: If8e2e8cc3fca855bd6b14eb3a957dadddfe143ed