doc: disable intersphinx mapping
The intersphinx extensions let one reserve a link namespace and map it
to a remote URL, for example to refer to the Python documentation. A
drawback is that, from time to time, Sphinx will attempt to download an
object inventory from the remote, that is not convenient when running
sphinx in offline mode such as in a plane.

Change-Id: Idef8e88545715a3f1cb9ace002960a7ad3e37b69