register reporters on test config
When doing the layout validation, I ended up spammed with a few errors:

  ERROR:zuul.Scheduler:Invalid reporter name gerrit

The issue is that my pipelines use 'gerrit' as reporter while it is not
registered when testing the config.  I registered the 'smtp' reporter as
well and the error is gone.

Side effect: the layout validation output now dumps actions for
start/success/failure.

Change-Id: I271a2943fa3e846ae60d9b615cd3a1ac3815bb1b