Make all changes reportable
We want to report on "changes" triggered by a timer.  Since those
aren't really changes, they are represented by a class called
NullChange which carries as much of the interface for a Change
object as possible (not much) so that they can be enqueued in
pipelines.

If a NullChange is reportable, then pretty much any kind of
change should be considered reportable.  So remove the flag that
indicates that NullChanges and Refs are not reportable.

Only attempt to format a change report if there is an action
defined for that pipeline (in case the default formatting process
attempts to access a change attribute that is inappropriate).
Remove checks that try to avoid formatting or sending a report
based on attributes of the change (which are no longer relevant).

Add a test for using a timer trigger with an smtp reporter.

Add validation of the attributes that an smtp reporter can use in
the layout file.

Allow the operator to configure a Subject for smtp reports.

Change-Id: Icd067a7600c2922a318b9ade11b3946df4e53065