make enqueue_time passable to addChange
When doing a doPromote we should keep the enqueue times of items.
However, the teardown and rebuild of the queues means that items
are fully destroyed and created a new.

Change the addChange api call so that it takes an optional enqueue
time to set on the item, which it can do internally once the new
item has gotten created.

This should address the issue where enqueue_times get lost when
we do a promote of the queue.

Now with unit tests! (Also ensured that if I removed the scheduler
change these tests failed, so the test is testing a correct thing)

Change-Id: I4dce23d00128280c7add922ca9ef5018b57d1cf3