Don't store change_queue in QueueItem
This was causing a problem with window sizes on reconfiguration because
the ChangeQueue objects were persisting across the reload via the local
reference inside of QueueItem.  Instead of adding more complexity to
reset those on reEnqueue, drop that and instead find the change queue
via the change's project when needed.

Also fix the fact that the QueueItem pipeline reference was not being
updated (it was set to None before a re-enqueue but then not set to
the new pipeline value).

Change-Id: I7f7050bfec985972ad7a1bc89da02d7b0753b798