Fix ZUUL_COMMIT in certain conditions
In a dependent pipeline with changes to multiple branches of
a single project, ZUUL_COMMIT would sometimes point to a commit
further back in the series rather than the commit that should
be tested (ie, not the commit pointed to by ZUUL_REF).  It's
possible this situation could appear in other circumstances as
well.

The error was the re-use of a variable when creating Zuul refs
for the entire series.  The variable should have been reserved
to indicate the final commit but instead could end up containing
the most recent commit of another branch or even project.

A test case is updated to detect both the condition where the
ZUUL_COMMIT for two different changes are the same, as well as
when ZUUL_REF does not point to ZUUL_COMMIT.

Change-Id: I9074746070edeebdff9f2ef3c1bb00526c21cfee