Add de-duplication to periodic triggers
Don't enqueue more than one entry for a project in a pipeline
that is triggered by a timer.  Match on the project name as that
is the only distinguishing attribute for a NullChange at the moment
(adding branch could be a future enhancement if we add multi-branch
support to periodic triggers).

Change-Id: I5b723497b23077652f2fc64b0c5367a20e3174c2