Drop formatStatusHTML() methods
The HTML build-in /status page used to report a nicely crafted HTML
report of the jobs being run.  It has been removed when switching to
Gearman (1f4c2bb1046).  We now provide a /status.json and an example
webapp to format the json.

The formatStatusHTML() methods did not get cleaned up.

Change-Id: I3051e81894461fc95d7cc238df1139b939fe01a2