Set Last-Modified header when serving status.json
Apache won't cache anything without a Last-Modified header. Be a good
citizen and set the Last-Modified header when serving status.json data.
Set the value to the cache_time timestamp.

Update the status checking test to make sure that this and other headers
are properly present so that this does not regress.

Change-Id: I2215d9c9694bd9ef7b94f29f654da77ae85a9959