Cause the promote command to ignore pipeline reqs
Because adding a change to a pipeline may, as a side effect, cause
the requirements for that pipeline not to be satisfied, ignore
pipeline requirements when changes are being promoted.  In this
case we are essentially trusting that the requirements were
satisfied during the initial enqueing and so should not be checked
again.

Change-Id: Ic86f3fb881ccd802a62a8f2f0cfb3ea5cccccb7b