Cause the enqueue command to ignore pipeline reqs
The enqueue command typically is used after a Zuul restart.  It
may also be useful when the normal method of triggering a change
is not possible, or perhaps too time consuming.  In all of those
cases, it may be desirable to bypass pipeline requirements (and
in the first case, it may be necessary since the act of enqueing
a change in a pipeline may cause it, as a side effect, to no longer
satisfy the requirements for that pipeline).

To resolve, this, ignore the pipeline requirements when a change
is enqueued into a pipeline as the result of an administrative
enqueue command.

Change-Id: Ic672452cd99355e3ee91df3c649f944ef75958a9