Fix queries for Zuul trigger
Against a live gerrit, we observed that sometimes gerrit may not
return data for a change.  In the case of the zuul trigger's
project-change-merged event, it's best to just ignore errors
for those changes and proceed.

Also, the gerrit trigger was attempting to parse the timing
results from the query.  The test did not catch this because it
did not supply them in the mocked method.  Correct this as well.

Also, double check that the query used was the one expected in
the test.

Change-Id: I792127d29f67f53a419eb94e9e0afb83b6e1bcb2