Stop removing changes from the cache
The zuul trigger ends up querying a large number of changes on
each merge, most of which are not in any pipelines.  To avoid
needlessly querying gerrit for data that do not change, keep
all changes that Zuul ever sees in the cache.

We could consider removing changes that are both not in pipelines
and closed (ie, keep all open changes), but we still end up querying
a number of merged changes each time due to Zuul following
dependencies of open changes.

Change-Id: Ie78df9aa43ec5ac35bdea79dcdafdfdd41d51d5b