Fix cloner test
The cloner test was incorrectly asserting that project2 should
always be at the commit specified for change B.  However, when
testing change A, project2's repo should be at master.

The cloner actually works as intended, the reason the assertion
passed is that the cloner was being run _after_ the gate jobs
finished, meaning that the master branch of project2 actually
was the commit specified for change B -- because it had merged.

Correct this by moving the build release to the end of the test,
so that the test operates entirely within the time that the jobs
are building (before anything merges).  Change the checks that
the commits are correct so that they assert the different and
correct values for each of the builds being tested.

Additionally, remove some unecessary actions from the test (adding
additional patchsets and specifying a clone map that is not used).

Change-Id: I7575e8b84925115393cef81d6b02a3532913ffed