Handle random PYTHONHASHSEED values.
Voluptions does a depth first in order traversal of the data structures
it is checking and fails fast. This means it will die when it hits the
first error and return only that error. The test_layouts test is
checking that different bad layout files fail in unique ways. This check
breaks when the python hashseed is random resulting in different depth
first traversals which can result in returning the same error twice if
bad layout files have common errors.

Fix this by trimming down the errors in each layout file to a single
error this way the order of the errors hit never matters.

Fix a string typo in test_layouts while we are in there.

Change-Id: Ie7587ebac4d22063a6819c68011f7fb67a49fcc0